Finalize postcode data cleaning

+++ src/main/java/com/poorna/australianPostCodes/Repository/AustraliaPostCodeRepository.java
@@ -3,5 +3,5 @@ package com.poorna.australianPostCodes.Repository;
 import com.poorna.australianPostCodes.Entity.AustralianPostCode;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AustraliaPostCodeRepository extends CrudRepository<AustralianPostCode, Integer> {
+public interface AustraliaPostCodeRepository extends CrudRepository<AustralianPostCode, Long> {
 }
